Implementation refactored into 2 InfoContributors
