+++ sip-core/src/main/java/de/ikor/sip/foundation/core/actuator/routes/MarkdownFilesContributor.java
@@ -11,22 +11,16 @@ import org.springframework.stereotype.Component;
 import org.springframework.util.FileCopyUtils;
 
 /**
- * {@link DetailsInfoContributor} extends {@link InfoContributor} to add needed information to the
+ * {@link MarkdownFilesContributor} extends {@link InfoContributor} to add needed information to the
  * info actuator
  */
 @Slf4j
 @Component
-public class DetailsInfoContributor implements InfoContributor {
+public class MarkdownFilesContributor implements InfoContributor {
 
   private static final String MARKDOWN_EXTENSION = ".md";
   private static final String BUILD_KEY = "build";
   private static final String FILES_KEY = "files";
-  private static final String ADAPTER_NAME_DETAILS_KEY = "adapter-name";
-  private static final String ADAPTER_NAME_BUILD_KEY = "name";
-  private static final String ADAPTER_VERSION_DETAILS_KEY = "adapter-version";
-  private static final String ADAPTER_VERSION_BUILD_KEY = "version";
-  private static final String SIP_FRAMEWORK_VERSION_DETAILS_KEY = "sip-framework-version";
-  private static final String SIP_FRAMEWORK_VERSION_BUILD_KEY = "sipFrameworkVersion";
 
   /**
    * Adding Markdown files {@link MarkdownObject} and base project information in actuator/info
@@ -39,30 +33,11 @@ public class DetailsInfoContributor implements InfoContributor {
   public void contribute(Info.Builder builder) {
     Map<String, Object> buildInfo = (LinkedHashMap<String, Object>) builder.build().get(BUILD_KEY);
 
-    if (buildInfo != null) {
-      collectAdapterInfo(buildInfo);
-    } else {
+    if (buildInfo == null) {
       buildInfo = new LinkedHashMap<>();
       builder.withDetail(BUILD_KEY, buildInfo);
     }
-
     buildInfo.put(FILES_KEY, fetchMarkdownObjects());
-  }
-
-  private void collectAdapterInfo(Map<String, Object> buildInfo) {
-    String adapterName = (String) buildInfo.get(ADAPTER_NAME_BUILD_KEY);
-    String adapterVersion = (String) buildInfo.get(ADAPTER_VERSION_BUILD_KEY);
-    String sipFrameworkVersion = (String) buildInfo.get(SIP_FRAMEWORK_VERSION_BUILD_KEY);
-
-    clearOriginalBuildInfo(buildInfo);
-
-    buildInfo.put(ADAPTER_NAME_DETAILS_KEY, adapterName);
-    buildInfo.put(ADAPTER_VERSION_DETAILS_KEY, adapterVersion);
-    buildInfo.put(SIP_FRAMEWORK_VERSION_DETAILS_KEY, sipFrameworkVersion);
-  }
-
-  private void clearOriginalBuildInfo(Map<String, Object> buildInfo) {
-    buildInfo.clear();
   }
 
   private List<MarkdownObject> fetchMarkdownObjects() {
